Data for the stroke prevention NMA-network discussed in Cooper et al. (2009)

The stroke prevention dataset describes a network comparing 17 treatments with 26 trials.

CooperStrokedat

Format

A data frame with 26 observations on the following 17 variables.

r1

Number of events (responses) in the baseline treatment (treatment 1).

r2

Number of events (responses) in the second study arm treatment (treatment 2).

r3

Number of events (responses) in the third study arm treatment (treatment 3).

r4

Number of events (responses) in the third study arm treatment (treatment 4).

r5

Number of events (responses) in the third study arm treatment (treatment 5).

n1

Total number of study participants in the baseline treatment (treatment 1).

n2

Total number of study participants in the second study arm treatment (treatment 2).

n3

Total number of study participants in the second study arm treatment (treatment 3).

n4

Total number of study participants in the second study arm treatment (treatment 4).

n5

Total number of study participants in the second study arm treatment (treatment 5).

t1

Indicator variable identifying treatment 1.

t2

Indicator variable identifying treatment 2.

t3

Indicator variable identifying treatment 3.

t4

Indicator variable identifying treatment 4.

t5

Indicator variable identifying treatment 5.

na

Indicator with number of arms in trial.

cov

Covariate, study-specific covariates

des

Design, the set of treatments included in each study

Source

Cooper, N.J., Sutton, A.J., Morris, D., Ades, A.E. and Welton, N.J., 2009. Addressing between-study heterogeneity and inconsistency in mixed treatment comparisons: Application to stroke prevention treatments in individuals with non-rheumatic atrial fibrillation. Statistics in medicine, 28(14), pp.1861-1881.