Fitting a pairwise meta-analysis model using INLA.

meta_inla fits a pairwise meta-analysis model using INLA

meta_inla(datINLA, fixed.par = c(0, 1000), tau.prior = "uniform",
  tau.par = c(0, 5), type = "FE", approach = "arm-level", mreg = FALSE,
  verbose = FALSE, inla.strategy = "simplified.laplace",
  improve.hyperpar.dz = 0.75, correct = FALSE, correct.factor = 10)

Arguments

datINLA
An object of create_INLA_dat_pair

fixed.par
A numerical vector specifying the parameter of the normal prior density for mean treatment effect, first value is parameter for mean, second is for variance.

tau.prior
A string specifying the prior density for the heterogeneity standard deviation, options are 'uniform' for uniform prior and 'half-normal' for half-normal prior.

tau.par
A numerical vector specifying the parameter of the prior density for heterogenety stdev.

var.par = c(u, l): u is lower bound and l is upper bound when var.prior = 'uniform'

var.par = c(m, v): m is mean and v is variance when var.prior = 'uniform'

type
A string indicating the type of the model, options are "FE", "RE".

approach
A string indicating the approach of the model, options are "summary-level", "arm-level"

mreg
Logical indicating whether covariate(s) should be incorporated to fit a meta-regression model, default FALSE

verbose
Logical indicating whether the program should run in a verbose model, default FALSE.

inla.strategy
A string specfying the strategy to use for the approximations of INLA; one of 'gaussian', 'simplified.laplace' (default) or 'laplace', see ?INLA::control.inla.

improve.hyperpar.dz
Step length in the standardized scale used in the construction of the grid, default 0.75, see INLA::inla.hyperpar.

correct
Logical Add correction for the Laplace approximation, default FALSE, see INLA::inla.hyperpar.

correct.factor
Numerical Factor used in adjusting the correction factor if correct=TRUE, default 10, see INLA::inla.hyperpar.

Value

meta_inla returns a meta_inla object with components:

Details

The following model types are supported

FE, fixed-effect model

RE, random effects model

Examples

data('TBdat')
## Create the dataset suitable for INLA
TBdatINLA <- create_INLA_dat_pair(TBdat$TRT, TBdat$CON, TBdat$TRTTB, TBdat$CONTB)

## Fitting a random-effects model using arm-level approach
# NOT RUN {
if(requireNamespace('INLA', quietly = TRUE)){
 require('INLA', quietly = TRUE)
fit.TB.RE.INLA <- meta_inla(TBdatINLA, type = 'RE', approach = 'arm-level',
tau.prior = 'uniform', tau.par = c(0, 5))
}
# }